Drop support for several platforms
We only target .NET 8, and that is not supported on NetBSD, OpenBSD or
Solaris, so drop those platforms.
