+++ MetaBrainz.MusicBrainz.DiscId/Platforms/MacOS.cs
@@ -1,8 +1,8 @@
 namespace MetaBrainz.MusicBrainz.DiscId.Platforms;
 
-internal sealed class Darwin : Unsupported {
+internal sealed class MacOS : Unsupported {
 
   // TODO: Port disc_darwin.c (looks a lot harder than the others)
-  // Also: No access to an OSX system.
+  // Also: No access to a macOS system.
 
 }
