--- MetaBrainz.MusicBrainz.DiscId/Platforms/NetBsd.cs
@@ -1,7 +1,0 @@
-namespace MetaBrainz.MusicBrainz.DiscId.Platforms;
-
-internal sealed class NetBsd : NetOrOpenBsd {
-
-  // Currently just a placeholder for any functionality specific to NetBSD
-
-}
