--- MetaBrainz.MusicBrainz.DiscId/Platforms/OpenBsd.cs
@@ -1,7 +1,0 @@
-namespace MetaBrainz.MusicBrainz.DiscId.Platforms;
-
-internal sealed class OpenBsd : NetOrOpenBsd {
-
-  // Currently just a placeholder for any functionality specific to OpenBSD
-
-}
