--- MetaBrainz.MusicBrainz.DiscId/Platforms/Solaris.cs
@@ -1,8 +1,0 @@
-namespace MetaBrainz.MusicBrainz.DiscId.Platforms;
-
-internal sealed class Solaris : Unsupported {
-
-  // TODO: Port disc_solaris.c
-  // Couldn't get Mono to work on Solaris (no vendor package, and build from sources failed), so this is not likely to get support 
-
-}
